Migrated to using update-notifier
The use of uname and apt-show-versions for detecting if a reboot was needed
proved to be too fragile. Started using the standard Debian/Ubuntu tools in
update-notifier-common package to determine how many pending updates and
security updates there are, as well as whether the computer needs to be
rebooted. The only minor issue with this approach is that reboot check will not
work until at least one kernel has been installed after the
update-notifier-common install.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>
Signed-off-by: Samuli Seppänen <samuli@openvpn.net>